+++ CHANGELOG.md
@@ -25,6 +25,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+## [9.5.0] 2023-04-12
+
+### Added
+
+- Added examples for time picker on `DatePickerInput`
+- Added `ToggleInput` component
+- Added `ToggleButton` component
+
 ## [9.4.0] 2023-04-11
 
 ### Added